requirements.txt
fastapi>=0.98.0
confluent-kafka>=2.1.1
uvicorn>=0.22.0
python-dotenv>=1.0.0